small amd64 tweaks

+++ metamod/api_hook.h
@@ -41,10 +41,10 @@
 #define _COMBINE2(x,y) x##y
 
 // simplified 'void' version of main hook function
-void DLLINTERNAL main_hook_function_void(unsigned long api_info_offset, enum_api_t api, unsigned long func_offset, const void * packed_args);
+void DLLINTERNAL main_hook_function_void(unsigned int api_info_offset, enum_api_t api, unsigned int func_offset, const void * packed_args);
 
 // full return typed version of main hook function
-void * DLLINTERNAL main_hook_function(const class_ret_t ret_init, unsigned long api_info_offset, enum_api_t api, unsigned long func_offset, const void * packed_args);
+void * DLLINTERNAL main_hook_function(const class_ret_t ret_init, unsigned int api_info_offset, enum_api_t api, unsigned int func_offset, const void * packed_args);
 
 //
 // API function args structures
